+++ book/guesser_test.go
@@ -23,7 +23,7 @@ func TestSeriesGuesser(t *testing.T) {
 		var got map[string]string
 
 		for _, re := range seriesGuessers {
-			got = submatchAsMap(tc.in, re)
+			got = reFindStringSubmatchAsMap(tc.in, re)
 			if got != nil {
 				break
 			}
@@ -51,11 +51,47 @@ func TestPathGuesser(t *testing.T) {
 		var got map[string]string
 
 		for _, re := range pathGuessers {
-			got = submatchAsMap(tc.in, re)
+			got = reFindStringSubmatchAsMap(tc.in, re)
 			if got != nil {
 				break
 			}
 		}
+		if fmt.Sprint(got) != fmt.Sprint(tc.out) {
+			t.Errorf("Guessing %#v failed:\nWant: %#v\nGot : %#v\n\n", tc.in, tc.out, got)
+		}
+	}
+}
+
+func TestContentGuesser(t *testing.T) {
+	testCases := []struct {
+		in  string
+		out map[string]string
+	}{
+		{in: `ISBN 978-0-596-52068-7`, out: map[string]string{"ISBN": "978-0-596-52068-7"}},
+		{in: `ISBN 978 0 596 52068 7`, out: map[string]string{"ISBN": "978 0 596 52068 7"}},
+		{in: `EAN: 9780596520687`, out: map[string]string{"ISBN": "9780596520687"}},
+		{in: `ISBN-10: 0-596-52068-9`, out: map[string]string{"ISBN": "0-596-52068-9"}},
+		{in: `ISBN : 978-2-7470-9059-9`, out: map[string]string{"ISBN": "978-2-7470-9059-9"}},
+		{in: `<br>ISBN 978 0 596 52068 7</br>`, out: map[string]string{"ISBN": "978 0 596 52068 7"}},
+		{in: "ISBN 99921-58-10-7", out: map[string]string{"ISBN": "99921-58-10-7"}},
+		{in: "ISBN 9971-5-0210-0", out: map[string]string{"ISBN": "9971-5-0210-0"}},
+		{in: "ISBN 960-425-059-0", out: map[string]string{"ISBN": "960-425-059-0"}},
+		{in: "ISBN 80-902734-1-6", out: map[string]string{"ISBN": "80-902734-1-6"}},
+		{in: "ISBN 85-359-0277-5", out: map[string]string{"ISBN": "85-359-0277-5"}},
+		{in: "ISBN 1-84356-028-3", out: map[string]string{"ISBN": "1-84356-028-3"}},
+		{in: "ISBN 0-684-84328-5", out: map[string]string{"ISBN": "0-684-84328-5"}},
+		{in: "ISBN 0-8044-2957-X", out: map[string]string{"ISBN": "0-8044-2957-X"}},
+		{in: "ISBN 0-85131-041-9", out: map[string]string{"ISBN": "0-85131-041-9"}},
+		{in: "ISBN 93-86954-21-4", out: map[string]string{"ISBN": "93-86954-21-4"}},
+		{in: "ISBN 0-943396-04-2", out: map[string]string{"ISBN": "0-943396-04-2"}},
+		{in: "ISBN 0-9752298-0-X", out: map[string]string{"ISBN": "0-9752298-0-X"}},
+		{in: `̀ISBN 12345`, out: nil},
+		{in: `11 – X A12616 ISBN 978-2-07-012616-3 13,90`, out: map[string]string{"ISBN": "978-2-07-012616-3"}},
+	}
+
+	for _, tc := range testCases {
+		got := reFindStringSubmatchAsMap(tc.in, contentGuesser)
+
 		if fmt.Sprint(got) != fmt.Sprint(tc.out) {
 			t.Errorf("Guessing %#v failed:\nWant: %#v\nGot : %#v\n\n", tc.in, tc.out, got)
 		}
@@ -82,7 +118,7 @@ func TestTitleCleaner(t *testing.T) {
 		var got map[string]string
 
 		for _, re := range titleCleaners {
-			got = submatchAsMap(tc.in, re)
+			got = reFindStringSubmatchAsMap(tc.in, re)
 			if got != nil {
 				break
 			}